Concrete-Crack-Detection-Segmentation/requirements.txt
# --- FastAPI and server dependencies ---
fastapi
uvicorn
pydantic
python-multipart
starlette

# --- CORS middleware ---
fastapi[all]  # Optional if you want all extras like CORS, Jinja2, etc.
# or just
# fastapi
# aiofiles  # if you serve static files

# --- Inference and image processing ---
numpy>=1.21.0
opencv-python>=4.5.0
Pillow
torch>=2.0.0
torchvision>=0.15.0
ultralytics
scikit-learn>=0.24.2
scikit-image>=0.18.0
albumentations>=1.1.0
pyyaml>=5.4.1

# --- Data analysis and visualization ---
pandas>=1.3.0
matplotlib>=3.4.3

# --- Logging and tracking ---
wandb>=0.12.0
tqdm>=4.62.0
imutils
requests
